fix(java): simplify cluster response handling and move cache test to ClientSideCacheTests
